refactor(server): folder structure
