+++ packages/amplication-server/src/core/dbSchemaImport/constants.ts
@@ -1,5 +1,5 @@
-import { Prisma } from "../../../prisma";
-import { EnumActionLogLevel, EnumActionStepStatus } from "../../action/dto";
+import { Prisma } from "../../prisma";
+import { EnumActionLogLevel, EnumActionStepStatus } from "../action/dto";
 
 export const PROCESSING_PRISMA_SCHEMA = "Processing Prisma schema";
