+++ packages/amplication-server/src/core/dbSchemaImport/dbSchemaImport.service.ts
@@ -1,20 +1,20 @@
 import { Inject, Injectable } from "@nestjs/common";
 import { KafkaProducerService } from "@amplication/util/nestjs/kafka";
 import { DBSchemaImportRequest } from "@amplication/schema-registry";
-import { User } from "../../../models";
-import { PrismaService } from "../../../prisma";
+import { User } from "../../models";
+import { PrismaService } from "../../prisma";
 import { ConfigService } from "@nestjs/config";
-import { Env } from "../../../env";
-import { EntityService, UserService } from "../..";
-import { EnumUserActionType } from "../types";
-import { AmplicationError } from "../../../errors/AmplicationError";
+import { Env } from "../../env";
+import { EntityService, UserService } from "..";
+import { EnumUserActionType } from "../userAction/types";
+import { AmplicationError } from "../../errors/AmplicationError";
 import { isDBImportMetadata } from "./utils/type-guards";
-import { UserAction } from "../dto";
+import { UserAction } from "../userAction/dto";
 import { PROCESSING_PRISMA_SCHEMA, initialStepData } from "./constants";
-import { ActionStep } from "../../action/dto";
-import { ActionService } from "../../action/action.service";
+import { ActionStep } from "../action/dto";
+import { ActionService } from "../action/action.service";
 import { AmplicationLogger } from "@amplication/util/nestjs/logging";
-import { UserActionService } from "../userAction.service";
+import { UserActionService } from "../userAction/userAction.service";
 import { DBSchemaImportMetadata } from "./types";
 import { CreateDBSchemaImportArgs } from "./dto/CreateDBSchemaImportArgs";
